+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added Function to Check the OS Type to make code in the script which runs only
 on a particular OS
 - added Function to get the default dirs
+- added getcwd and chdir function
 
 ## 0.1.9
 - made the deprecated warning yellow